+++ flash/image/instance_segmentation/backbones.py
@@ -14,12 +14,14 @@
 from functools import partial
 from typing import Optional
 
+from lightning_utilities.core.imports import module_available
+
 from flash.core.adapter import Adapter
 from flash.core.integrations.icevision.adapter import IceVisionAdapter, SimpleCOCOMetric
 from flash.core.integrations.icevision.backbones import get_backbones, load_icevision_ignore_image_size
 from flash.core.model import Task
 from flash.core.registry import FlashRegistry
-from flash.core.utilities.imports import _ICEVISION_AVAILABLE, _module_available, _TORCHVISION_AVAILABLE
+from flash.core.utilities.imports import _ICEVISION_AVAILABLE, _TORCHVISION_AVAILABLE
 from flash.core.utilities.providers import _ICEVISION, _MMDET, _TORCHVISION
 
 if _ICEVISION_AVAILABLE:
@@ -71,7 +73,7 @@ if _ICEVISION_AVAILABLE:
             providers=[_ICEVISION, _TORCHVISION],
         )
 
-    if _module_available("mmdet"):
+    if module_available("mmdet"):
         model_type = icevision_models.mmdet.mask_rcnn
         INSTANCE_SEGMENTATION_HEADS(
             partial(load_icevision_ignore_image_size, model_type),